fix: script broke on beta (hydration failed)
Script seems to interrupt SSR on beta (React 18),
resulting in the following error:
Hydration failed because the initial UI does not match
what was rendered on the server.
Added a delay to wait for SSR to complete before manipulating DOM.
Might also fix the conflicts with other scripts that
cause a right-alignment of track titles and failed initialization.